Avoid jitter
This was visible when drawing the same data at different
start/stop viewports. The rounding of coordinates during
the draw cycle causes the trace to shift around slightly,
making it look like the data is changing.

Solution is to calculate positions at a fixed offset in
integer coordinates. Then offset during the draw cycle
to the actual start position (also in integer coords).